requirements-convert.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.1
torchvision==0.18.1
transformers==4.41.2
optimum[openvino]==1.20.0
openvino-dev==2024.1.0